requirements.txt
plotly
matplotlib
kaleido
pandas
pyyaml
networkx